feat: serverspec (#1)
## The Issue

This pull request introduces Serverspec integration into the DDEV
environment, enabling users to perform infrastructure testing within
their projects.

This is the initial implementation of the service.

<!-- Provide a brief description of the issue. -->

## How This PR Solves The Issue

<!-- Describe the key change(s) in this PR that address the issue above.
-->

## Manual Testing Instructions

1.  **Install the add-on:**
    ```bash
ddev add-on get
https://github.com/tyler36/ddev-serverspec/tarball/refs/pull/1/head
    ddev restart
    ```

2.  **Initialize Serverspec tests:**
    ```bash
    ddev serverspec-init
    ```
Follow the prompts in the wizard, selecting options suitable for your
project. For example, select `1` (UN*X) and `2` (Exec (local)).

3.  **Run Serverspec tests:**
    ```bash
    ddev serverspec
    ```
Verify that the tests run successfully and output something similar to
`1 example, 0 failures`.

4. **Add tests for your application.** For example, you might test that
the web container listens on port 80/443, or that necessary php
extensions are installed.


## Automated Testing Overview

<!-- Please describe the tests introduced by this PR, or explain why no
tests are needed. -->

## Release/Deployment Notes

<!-- Does this affect anything else or have ramifications for other
code? Does anything have to be done on deployment? -->