refactor: rename `serverspec` -> `rspec` (#3)
## The Issue

This pull request addresses an issue where the `ddev serverspec` command
was incorrectly configured to use `rake spec` to run ServerSpec tests,
which is deprecated. It updates the command to directly use `rspec`,
which is the more modern and recommended approach. This also involved
renaming the command file for accuracy.

## How This PR Solves The Issue

This PR updates its internals to use `rspec` instead of `rake`.
The command helper was renamed to `ddev rspec` to better reflect its
purpose (and it's shorter!).

## Manual Testing Instructions

<!-- If this PR changes logic, consider adding additional steps or
context to the instructions below. -->

1. Install PR #3

```bash
ddev add-on get https://github.com/tyler36/ddev-serverspec/tarball/refs/pull/3/head
ddev restart
```

2. Run test (or run `ddev serverspec-init` first if required) .

```
ddev rspec
```

## Automated Testing Overview

<!-- Please describe the tests introduced by this PR, or explain why no
tests are needed. -->

## Release/Deployment Notes

<!-- Does this affect anything else or have ramifications for other
code? Does anything have to be done on deployment? -->